+++ changelog.md
@@ -7,6 +7,7 @@ Changelog
 * Auto-generate readme with oclif.
 * Added support for registry:delete command, to delete events from the
   registry.
+* Added registry:list-entities command.
 
 
 0.1.0 (2023-03-29)